fix lint issues

+++ src/StakingNodesManager.sol
@@ -539,7 +539,7 @@ contract StakingNodesManager is
     function processPrincipalWithdrawals(
         WithdrawalAction[] memory actions
     ) public onlyRole(WITHDRAWAL_MANAGER_ROLE)  {
-        for (uint i = 0; i < actions.length; i++) {
+        for (uint256 i = 0; i < actions.length; i++) {
             _processPrincipalWithdrawalForNode(actions[i]);
         }
     }
