+++ src/interfaces/IStakingNodesManager.sol
@@ -17,12 +17,12 @@ interface IStakingNodesManager {
         bytes publicKey;
         bytes signature;
         bytes32 depositDataRoot;
-        uint nodeId;
+        uint256 nodeId;
     }
 
     struct Validator {
         bytes publicKey;
-        uint nodeId;
+        uint256 nodeId;
     }
 
     struct WithdrawalAction {
@@ -40,11 +40,11 @@ interface IStakingNodesManager {
     function getAllNodes() external view returns (IStakingNode[] memory);
     function isStakingNodesOperator(address) external view returns (bool);
     function isStakingNodesDelegator(address _address) external view returns (bool);
-    function processRewards(uint nodeId, RewardsType rewardsType) external payable;
+    function processRewards(uint256 nodeId, RewardsType rewardsType) external payable;
     function registerValidators(
         ValidatorData[] calldata _depositData
     ) external;
-    function nodesLength() external view returns (uint);
+    function nodesLength() external view returns (uint256);
 
     function upgradeableBeacon() external returns (UpgradeableBeacon);
 
@@ -58,5 +58,3 @@ interface IStakingNodesManager {
 
     function isStakingNodesWithdrawer(address _address) external view returns (bool);
 }
-
-
